refactor(generator): extract DtoFinder from SymbolDiscovery
